requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
google-genai>=1.0.0          # SDK unificado Gemini + Vertex AI
google-auth>=2.0.0

# ── RAG / Vector Store ───────────────────────────────────────────────────────
chromadb>=0.5.0              # Vector DB local (swap: Vertex AI Vector Search)

# ── Tools ────────────────────────────────────────────────────────────────────
requests>=2.31.0
pytrends>=4.9.2              # Google Trends scraping (local, sin API key)

# ── Utils ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0         # Carga .env en desarrollo local
numpy>=1.26.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0

# ── Opcional: LangChain integration ─────────────────────────────────────────
# langchain>=0.3.0
# langchain-google-genai>=2.0.0

# ── Opcional: Streamlit UI ───────────────────────────────────────────────────
# streamlit>=1.35.0
